+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Unreleased
 - API changes:
     - WOM-479: Removed redundant ``get_enum_strings()`` method. Use ``get_attribute_data_type()`` instead.
 - Non-breaking changes:
-    - WOM-386: Added new ``StaticPointingModel`` class for the import/export of a global static pointing model from/to a JSON file.
+    - WOM-386, WOM-445: Added new ``StaticPointingModel`` class for the import/export of a global static pointing model from/to a JSON file.
     - WOM-479: Updated ``get_attribute_data_type()`` to fix exceptions in DiSQ.
     - WOM-484: Updated ``subscribe()`` to subscribe to all input nodes in one OPCUA call.
     - WOM-446, WOM-464: Added missing commands to the ``Command`` enum.